RDM-6520: Validate the CaseRoles on the case submission
